reshuffle setCmdUser
